Create Sudoku Solver
Input Format: 
The program requires input of the Sudoku board in a 9x9 format, where any empty cell is represented by 0.

Run the Program: 
Compile and run the code. You'll be prompted to enter the Sudoku board.

Note:
This program implements a backtracking algorithm to solve the Sudoku puzzle.
The program will display colored outputs for better visual appeal.